checkers/kubik/requirements.txt
checklib==0.7.0
paho-mqtt==2.0.0
requests==2.31.0